+++ src/Control/Eff/Concurrent/Protocol/Request.hs
@@ -4,7 +4,12 @@
 module Control.Eff.Concurrent.Protocol.Request
   ( Request(..)
   , sendReply
-  , sendEmbeddedReply
+  , ReplyTarget(..)
+  , replyTarget
+  , replyTargetOrigin
+  , replyTargetSerializer
+  , embeddedReplyTarget
+  , toEmbeddedReplyTarget
   , RequestOrigin(..)
   , embedRequestOrigin
   , toEmbeddedOrigin
@@ -18,10 +23,10 @@ import Control.DeepSeq
 import Control.Eff
 import Control.Eff.Concurrent.Process
 import Control.Eff.Concurrent.Protocol
-import Data.Functor.Contravariant
+import Control.Lens
 import Data.Kind (Type)
 import Data.Typeable (Typeable)
-import Data.Tagged
+import Data.Semigroup
 import GHC.Generics
 
 -- | A wrapper sum type for calls and casts for the 'Pdu's of a protocol
@@ -95,37 +100,6 @@ makeRequestOrigin = RequestOrigin <$> self <*> makeReference
 
 instance NFData (RequestOrigin p r)
 
--- | Send a 'Reply' to a 'Call'.
---
--- The reply will be deeply evaluated to 'rnf'.
---
--- To send replies for 'EmbedProtocol' instances use 'sendEmbeddedReply',
--- or 'embedReplySerializer' and 'toEmbeddedOrigin'.
---
--- @since 0.15.0
-sendReply ::
-  forall protocol reply q eff .
-     (SetMember Process (Process q) eff, Member Interrupts eff, Tangible reply, Typeable protocol)
-  => Serializer (Reply protocol reply) -> RequestOrigin protocol reply -> reply -> Eff eff ()
-sendReply ser o r = sendAnyMessage (_requestOriginPid o) $! runSerializer ser $! Reply o r
-
--- | Send a 'Reply' to a 'Call' for an embedded 'Pdu'.
---
--- The reply will be deeply evaluated to 'rnf'.
---
--- @since 0.25.1
-sendEmbeddedReply ::
-  forall outer inner reply q eff .
-     ( SetMember Process (Process q) eff
-     , Member Interrupts eff
-     , TangiblePdu outer ('Synchronous reply)
-     , TangiblePdu inner ('Synchronous reply)
-     , Tangible reply
-     , EmbedProtocol outer inner
-     )
-  => Serializer (Reply outer reply) -> RequestOrigin outer reply -> Tagged inner reply -> Eff eff ()
-sendEmbeddedReply ser o = sendReply @outer @reply (embedReplySerializer ser) (toEmbeddedOrigin o) . unTagged
-
 -- | Turn an 'RequestOrigin' to an origin for an embedded request (See 'EmbedProtocol').
 --
 -- This is useful of a server delegates the @calls@ and @casts@ for an embedded protocol
@@ -171,3 +145,81 @@ embedReplySerializer = contramap embedReply
 -- @since 0.24.2
 embedReply :: forall outer inner reply . EmbedProtocol outer inner => Reply inner reply -> Reply outer reply
 embedReply (Reply (RequestOrigin !pid !ref) !v) = Reply (RequestOrigin pid ref) v
+
+
+-- | Answer a 'Call' by sending the reply value to the client process.
+--
+-- The 'ProcessId', the 'RequestOrigin' and the 'Reply' 'Serializer' are
+-- stored in the 'ReplyTarget'.
+--
+-- @since 0.25.1
+sendReply
+  :: ( SetMember Process (Process q) eff
+     , Member Interrupts eff
+     , Tangible reply
+     , Typeable protocol
+     )
+  => ReplyTarget protocol reply
+  -> reply
+  -> Eff eff ()
+sendReply (MkReplyTarget (Arg o ser)) r =
+  sendAnyMessage (_requestOriginPid o) $! runSerializer ser $! Reply o r
+
+-- | Target of a 'Call' reply.
+--
+-- This combines a 'RequestOrigin' with a 'Serializer' for a 'Reply' using 'Arg'.
+-- There are to smart constructors for this type: 'replyTarget' and 'embeddedReplyTarget'.
+--
+-- Because of 'Arg' the 'Eq' and 'Ord' instances are implemented via
+-- the 'RequestOrigin' instances.
+--
+-- @since 0.26.0
+newtype ReplyTarget p r =
+  MkReplyTarget (Arg (RequestOrigin p r) (Serializer (Reply p r)))
+    deriving (Eq, Ord, Typeable)
+
+instance Show (ReplyTarget p r) where
+  showsPrec d (MkReplyTarget (Arg o _s)) = showParen (d>=10) (showString "reply-target: " . shows o)
+
+instance NFData (ReplyTarget p r) where
+  rnf (MkReplyTarget (Arg x y)) = rnf x `seq` y `seq` ()
+
+-- | Smart constructor for a 'ReplyTarget'.
+--
+-- To build a @ReplyTarget@ for an 'EmbedProtocol' instance use 'embeddedReplyTarget'.
+--
+-- @since 0.26.0
+replyTarget :: Serializer (Reply p reply) -> RequestOrigin p reply -> ReplyTarget p reply
+replyTarget ser orig =  MkReplyTarget (Arg orig ser)
+
+-- | A simple 'Lens' for the 'RequestOrigin' of a 'ReplyTarget'.
+--
+-- @since 0.26.0
+replyTargetOrigin :: Lens' (ReplyTarget p reply) (RequestOrigin p reply)
+replyTargetOrigin f (MkReplyTarget (Arg o x)) =
+  (\o' -> MkReplyTarget (Arg o' x)) <$> f o
+
+-- | A simple 'Lens' for the 'Reply' 'Serializer' of a 'ReplyTarget'.
+--
+-- @since 0.26.0
+replyTargetSerializer :: Lens' (ReplyTarget p reply) (Serializer (Reply p reply))
+replyTargetSerializer f (MkReplyTarget (Arg x o)) =
+  (\o' -> MkReplyTarget (Arg x o')) <$> f o
+
+-- | Smart constructor for an /embedded/ 'ReplyTarget'.
+--
+-- This combines 'replyTarget' and 'toEmbeddedReplyTarget'.
+--
+-- @since 0.26.0
+embeddedReplyTarget :: EmbedProtocol outer inner => Serializer (Reply outer reply) -> RequestOrigin outer reply -> ReplyTarget inner reply
+embeddedReplyTarget ser orig = toEmbeddedReplyTarget $ replyTarget ser orig
+
+-- | Convert a 'ReplyTarget' to be usable for /embedded/ replies.
+--
+-- This combines a 'toEmbeddedOrigin' with 'embedReplySerializer' to produce a
+-- 'ReplyTarget' that can be passed to functions defined soley on an embedded protocol.
+--
+-- @since 0.26.0
+toEmbeddedReplyTarget :: EmbedProtocol outer inner => ReplyTarget outer r -> ReplyTarget inner r
+toEmbeddedReplyTarget (MkReplyTarget (Arg orig ser)) =
+  MkReplyTarget (Arg (toEmbeddedOrigin orig) (embedReplySerializer ser))